# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Property Management System
  version: 1.0.0
  description: |
    Source of truth for the boundary between the Angular application and the Go service.
    The Go server interface and the Angular client are both generated from this document;
    neither side hand-writes these types (Constitution III).

    Authorization is decided on the server for every operation. `x-required-role` states
    the requirement; an operation marked `admin` is refused for a manager regardless of
    how the request is made (Constitution I). Configuration — property types, areas,
    custom field definitions, and changing a property's reference code — are `admin`.
    Everything else is `manager`, which an administrator also satisfies.

    Sensitive custom fields never appear in this document as a filterable field. The
    server refuses a filter naming one even if a client constructs the request by hand
    (FR-027s4, Constitution VII).

    Every response body uses Western Arabic digits. Messages shown to users are Arabic.
    All timestamps are RFC 3339 in UTC; the client renders them in Africa/Cairo.
servers:
  - url: https://localhost:8443/api/v1
    description: Local development. TLS only — there is no plaintext listener.
...
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      operationId: signIn
      summary: Sign in
      description: |
        Verifies a username and password and issues a session cookie.

        The refusal for an unknown username, a wrong password, and a deactivated
        account is byte-identical. Repeated failures impose a growing wait that
        applies equally to usernames that do not exist; while a wait is outstanding
        the attempt is refused as `too_soon` even if the password is correct.
      security: []
      x-required-role: none
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                password:
                  type: string
                  minLength: 1
                  maxLength: 256
      responses:
        '200':
          description: |
            Signed in. Sets the session cookie with `HttpOnly`, `Secure`,
            `SameSite=Strict`.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: 'pms_session=<opaque token>; HttpOnly; Secure; SameSite=Strict; Path=/'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          description: |
            Refused. The body is identical for an unknown username, a wrong password,
            and a deactivated account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Submitted before the current waiting period elapsed.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    required: [retryAfterSeconds]
                    properties:
                      retryAfterSeconds:
                        type: integer
                        minimum: 1
                        maximum: 60
                        description: Seconds remaining before another attempt is accepted.
  /auth/logout:
    post:
      tags: [auth]
      operationId: signOut
      summary: Sign out
      description: Revokes the current session immediately.
      x-required-role: any
      responses:
        '204':
          description: Signed out; the session cookie is cleared and the session revoked.
        '401':
          $ref: '#/components/responses/Error'
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: The signed-in account as it stands right now
      description: |
        Reads role and active state from stored data on each call, so a demotion or
        deactivation is reflected on the next request without signing out.
      x-required-role: any
      responses:
        '200':
          description: The caller's current identity, role, and password-change state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Error'
  /auth/password:
    post:
      tags: [auth]
      operationId: changeOwnPassword
      summary: Change your own password
      description: |
        Requires the current password. On success every session belonging to the
        account is revoked, including this one, and the caller must sign in again.
      x-required-role: any
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  maxLength: 256
                newPassword:
                  type: string
                  minLength: 12
                  maxLength: 256
      responses:
        '204':
          description: Password changed; all sessions revoked.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: List accounts
      x-required-role: admin
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: includeInactive
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: A page of accounts, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items, page, pageSize, totalItems]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  page: { type: integer }
                  pageSize: { type: integer }
                  totalItems: { type: integer }
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      tags: [users]
      operationId: createUser
      summary: Create an account
      description: |
        There is no self-registration; only an administrator reaches this operation.
        The new account must change its password at first sign-in. A username that
        canonicalizes to an existing one is refused as already in use.
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, displayName, role, initialPassword]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 120
                role:
                  $ref: '#/components/schemas/Role'
                initialPassword:
                  type: string
                  minLength: 12
                  maxLength: 256
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          description: A username that renders identically already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [users]
      operationId: getUser
      summary: Read one account
      x-required-role: admin
      responses:
        '200':
          description: The account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [users]
      operationId: updateUser
      summary: Change display name, role, or active state
      description: |
        A role change takes effect on the affected user's next request and does not end
        their session. Deactivation revokes every session belonging to the account.

        Refused when the change would leave no active administrator, and when an
        administrator targets their own account for demotion or deactivation.
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 120
                role:
                  $ref: '#/components/schemas/Role'
                isActive:
                  type: boolean
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409':
          description: |
            Refused because it would leave no active administrator, or because the
            caller targeted their own account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/password:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      tags: [users]
      operationId: resetUserPassword
      summary: Reset another account's password
      description: |
        Administrator only. The account must change the password at its next sign-in,
        and all its sessions are revoked. The current password is not required and is
        never readable.
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword:
                  type: string
                  minLength: 12
                  maxLength: 256
      responses:
        '204':
          description: Password reset; sessions revoked; change required at next sign-in.
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /audit-records:
    get:
      tags: [records]
      operationId: listAuditRecords
      summary: Read recorded actions
      description: |
        Administrator only. Newest first, paged, filters combinable. There is no
        operation anywhere in this document that edits or deletes a record.
      x-required-role: admin
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: actorId
          in: query
          description: Only actions performed by this account
          schema:
            type: string
            format: uuid
        - name: targetId
          in: query
          description: Only actions that affected this account
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          description: Only this kind of action
          schema:
            $ref: '#/components/schemas/AuditAction'
        - name: from
          in: query
          description: Start of the range, inclusive of the whole day in Africa/Cairo
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End of the range, inclusive of the whole day in Africa/Cairo
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A page of records, newest first. An empty page is a valid answer.
          content:
            application/json:
              schema:
                type: object
                required: [items, page, pageSize, totalItems]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditRecord'
                  page: { type: integer }
                  pageSize: { type: integer }
                  totalItems: { type: integer }
        '400':
          description: Invalid filter — for example a start date later than the end date.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /properties:
    get:
      tags: [properties]
      operationId: listProperties
      summary: A page of the register
      description: |
        Sorted by name by default. `q` matches the name and the reference code after
        Arabic normalisation. `code` and `name` are per-column filters matching those
        single columns. Every filter combines by AND. Archived properties are excluded
        unless `includeArchived` is true.
      x-required-role: manager
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, enum: [10, 25, 50, 100], default: 25 }
        - name: q
          in: query
          description: Matched against name and reference code.
          schema: { type: string, maxLength: 120 }
        - name: code
          in: query
          description: |
            Column filter: substring match against the reference code after Arabic
            normalisation. Combines with every other filter by AND.
          schema: { type: string, maxLength: 32 }
        - name: name
          in: query
          description: |
            Column filter: substring match against the property name after Arabic
            normalisation. Combines with every other filter by AND.
          schema: { type: string, maxLength: 120 }
        - name: propertyTypeId
          in: query
          schema: { type: string, format: uuid }
        - name: areaId
          in: query
          schema: { type: string, format: uuid }
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: A page of properties. An empty page is a valid answer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyPage' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }

    post:
      tags: [properties]
      operationId: createProperty
      summary: Add a property
      description: |
        `code` may be supplied only by an administrator; a manager supplying it is
        refused with 403. When omitted, the server generates the next code.
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyCreate' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409':
          description: The supplied reference code is already held, active or archived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /properties/search:
    post:
      tags: [search]
      operationId: searchProperties
      summary: Advanced search across built-in and custom fields
      description: |
        Filters combine with AND; an unset filter constrains nothing.

        A filter naming a sensitive field is refused with 400, not silently dropped
        (Constitution VII). The server resolves every `fieldId` against the definitions
        before building the query; an unknown id is refused.
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdvancedSearch' }
      responses:
        '200':
          description: A page of matching properties.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyPage' }
        '400':
          description: |
            An unknown field, a sensitive field, or an operator that does not suit the
            field's type.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /properties/{propertyId}:
    parameters:
      - $ref: '#/components/parameters/PropertyId'
    get:
      tags: [properties]
      operationId: getProperty
      summary: One property with all of its fields
      x-required-role: manager
      responses:
        '200':
          description: The property, archived or not.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyDetail' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }

    put:
      tags: [properties]
      operationId: updateProperty
      summary: Modify a property
      description: |
        `version` must equal the version last read or the request is refused with
        `version_conflict`. The reference code is not part of this body; it has its own
        operation. An archived property is refused with `archived` until restored.
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyUpdate' }
      responses:
        '200':
          description: Updated, with the new version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyDetail' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409':
          description: |
            The property changed since it was read, or it is archived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /properties/{propertyId}/code:
    parameters:
      - $ref: '#/components/parameters/PropertyId'
    patch:
      tags: [properties]
      operationId: changePropertyCode
      summary: Set a property's reference code
      description: |
        Administrator only. A separate operation rather than a field in the update
        body, so that its authorization does not depend on which fields a request
        happens to carry (Constitution I).
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, version]
              properties:
                code:    { $ref: '#/components/schemas/PropertyCode' }
                version: { type: integer }
      responses:
        '200':
          description: The code was changed; the old and new codes are in the audit log.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyDetail' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /properties/{propertyId}/archive:
    parameters:
      - $ref: '#/components/parameters/PropertyId'
    post:
      tags: [properties]
      operationId: archiveProperty
      summary: Archive a property
      description: |
        The property leaves the working list but is never destroyed. Its reference
        code stays reserved forever.
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: integer }
      responses:
        '200':
          description: Archived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyDetail' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409':
          description: Already archived, or changed since it was read.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /properties/{propertyId}/restore:
    parameters:
      - $ref: '#/components/parameters/PropertyId'
    post:
      tags: [properties]
      operationId: restoreProperty
      summary: Restore an archived property
      x-required-role: manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: integer }
      responses:
        '200':
          description: Restored with every field as it was.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PropertyDetail' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /property-types:
    get:
      tags: [lookups]
      operationId: listPropertyTypes
      summary: Every property type
      description: Readable by anyone signed in — the property form needs it.
      x-required-role: manager
      responses:
        '200':
          description: All property types, ordered by label.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Lookup' }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [lookups]
      operationId: createPropertyType
      x-required-role: admin
      summary: Add a property type
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LookupWrite' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lookup' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /property-types/{lookupId}:
    parameters:
      - $ref: '#/components/parameters/LookupId'
    put:
      tags: [lookups]
      operationId: renamePropertyType
      summary: Rename a property type
      description: |
        Every property carrying it shows the new label immediately, because a property
        refers to the entry by identity.
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LookupWrite' }
      responses:
        '200':
          description: Renamed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lookup' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      tags: [lookups]
      operationId: deletePropertyType
      summary: Remove an unused property type
      x-required-role: admin
      responses:
        '204':
          description: Removed.
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409':
          description: |
            In use. The Arabic message states how many properties use it, archived
            included.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /areas:
    get:
      tags: [lookups]
      operationId: listAreas
      summary: Every area
      x-required-role: manager
      responses:
        '200':
          description: All areas, ordered by label.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Lookup' }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [lookups]
      operationId: createArea
      x-required-role: admin
      summary: Add an area
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LookupWrite' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lookup' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /areas/{lookupId}:
    parameters:
      - $ref: '#/components/parameters/LookupId'
    put:
      tags: [lookups]
      operationId: renameArea
      summary: Rename an area
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LookupWrite' }
      responses:
        '200':
          description: Renamed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lookup' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      tags: [lookups]
      operationId: deleteArea
      summary: Remove an unused area
      x-required-role: admin
      responses:
        '204':
          description: Removed.
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pms_session
      description: |
        Set by /auth/login. HttpOnly, Secure, SameSite=Strict. Not readable by
        JavaScript and never placed in a URL. Carries identity only; role and active
        state are read from stored data on every request.
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
...
  schemas:
...
    CurrentUser:
      type: object
      required: [id, username, displayName, role, mustChangePassword]
      description: |
        The caller as they stand on this request. `mustChangePassword` true means
        the client must send the user to the password-change screen and no other.
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        mustChangePassword: { type: boolean }
...